pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "djaodjin-pages>=0.8.3",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "plotly>=6.0.1",
    "requests>=2.32.3",
    "streamlit>=1.44.1",
    "tensorflow>=2.14.0",
]
